IMPROVE: move favoriteButton.tsx
